2019-08-27 Version 0.8.2: Updated CPU info

+++ app/src/main/java/com/smlnskgmail/jaman/deviceinfo/info/cpu/CPUInfo.kt
@@ -3,7 +3,7 @@ package com.smlnskgmail.jaman.deviceinfo.info.cpu
 interface CPUInfo {
 
     fun cores(): Int
-    fun minimumFreq(): Float
-    fun maximumFreq(): Float
+    fun minimumFreq(): Int
+    fun maximumFreq(): Int
 
 }
